+++ ContentPatcher/release-notes.md
@@ -1,6 +1,12 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.9.1
+Released 12 June 2019.
+
+* Fixed error loading local XNB files in some cases with Content Patcher 1.9.
+* Fixed mod-provided tokens being asked for values when they're marked non-ready.
+
 ## 1.9
 Released 09 June 2019.